refactor reporting batch pause route
