+++ src/app/router_registry.py
@@ -233,6 +233,7 @@ from app.routers.reporting_batch_controls import (
 from app.routers.reporting_batch_lease_recovery import (
     recovery_router as reporting_batch_lease_recovery_router,
 )
+from app.routers.reporting_batch_pause import pause_router as reporting_batch_pause_router
 from app.routers.reporting_batch_retry import retry_router as reporting_batch_retry_router
 from app.routers.reporting_batch_run_once import (
     worker_router as reporting_batch_run_once_router,
@@ -440,6 +441,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(reporting_batch_status_router)
     app.include_router(reporting_batch_controls_router)
     app.include_router(reporting_batch_cancel_router)
+    app.include_router(reporting_batch_pause_router)
     app.include_router(reporting_batch_retry_router)
     app.include_router(reporting_batch_lease_recovery_router)
     app.include_router(reporting_batch_run_once_router)
